+++ image.h
@@ -34,7 +34,7 @@ struct hash_box {Vector3d min, max; // axis aligned bounding box
 };
 struct HPoint {
 	Vector3d f,pos,nrm,flux; 
-	double r2; 
+	float r2; 
 	unsigned int n; // n = N / ALPHA in the paper
 	int pix;
 };
@@ -49,9 +49,9 @@ class Color{
         void print();
         friend ostream &operator<<(ostream&, const Color&);
         bool operator==(Color);
-        double operator-(Color);
+        float operator-(Color);
         Color operator+(Color);
-        Color operator*(double);
+        Color operator*(float);
 };
 Color get_color(string);
 class Vector3d;
@@ -63,16 +63,16 @@ static Vector3d Vector3d_static;
 struct Light{
     Vector3d from, direction, fin, handle;
     string color;
-    double range;
-    Light(Vector3d, Vector3d, double, string);
+    float range;
+    Light(Vector3d, Vector3d, float, string);
 };
 
 /* ------------------ Main Zone -------------*/
 
 struct Pair{
     int i;
-    double v;
-    Pair(int i=0, double v=0):i(i), v(v){}
+    float v;
+    Pair(int i=0, float v=0):i(i), v(v){}
 };
 class Image{
     public:
@@ -81,7 +81,7 @@ class Image{
         Color color;
         vector<unsigned char> img;
         vector<Vector3d> measurement;
-        vector<vector<double>> values;
+        vector<vector<float>> values;
         vector<vector<Pair>> DP_values;
         vector<int> seam;
         vector<vector<int>> seamsx;
@@ -91,10 +91,10 @@ class Image{
         void output(string outfile_name);
         void reset(int width=640, int height=480);
         bool from_file(string infile_name);
-        void draw_point(double =0, double =0, Color=get_color("white"));
+        void draw_point(float =0, float =0, Color=get_color("white"));
         void draw_point(int =0, int =0, Color=get_color("white"));
         Color pick_color(int, int);
-        Color pick_color(double, double);
+        Color pick_color(float, float);
         void draw_from_measurement();
         void calculate_value(string);
         void calculate_seam(string);
@@ -106,7 +106,7 @@ class Image{
         bool contains(string, int, int);
         void protect(int, int, int, int);
         void remove(int, int, int, int);
-        Image denoise(int size, double alpha);
+        Image denoise(int size, float alpha);
 };
 class Scene{
     public:
@@ -117,7 +117,7 @@ class Scene{
     long num_sample;
     int num_thread;
     int multi_select;
-    double scale;
+    float scale;
     Image image;
     vector<vector<string>> obj_file;
     vector<NurbsSurface> surfaces;
@@ -125,11 +125,11 @@ class Scene{
     BVH<Patch> bvh;
     vector<Patch*> patchList;
     unsigned int num_hash, num_photon;
-    double hash_s; List **hash_grid; List *hitpoints = NULL; hash_box hpbbox;
+    float hash_s; List **hash_grid; List *hitpoints = NULL; hash_box hpbbox;
 
     Scene(Vector3d from, Vector3d direction, vector<Light> lights,
             int width, int height,
-            double scale,
+            float scale,
             long num_sample,
             int multi_select, int num_thread);
     void sketch(NurbsCurve, int =200, Color =get_color("white"));
@@ -139,8 +139,8 @@ class Scene{
     void output(string outfile_name, int );
     void build_bvh();
     void render(string, int);
-    Color trace_ray(double, double);
-    Vector3d shoot_ray(double, double);
+    Color trace_ray(float, float);
+    Vector3d shoot_ray(float, float);
 
     inline unsigned int f_hash(const int ix, const int iy, const int iz);
     void build_hash_grid(const int w, const int h);
@@ -150,10 +150,10 @@ class Scene{
 
 struct bezier_coefficient_item{
     int n, k;
-    vector<vector<vector<double>>> coeffs;
+    vector<vector<vector<float>>> coeffs;
     bezier_coefficient_item(int =0,int =0);
 };
-Matrix Bezier(int, int, double);
+Matrix Bezier(int, int, float);
 class NurbsCurve{
     public:
     static int k_default, n_default;
@@ -162,13 +162,13 @@ class NurbsCurve{
     NurbsCurve(int n=1, int k=k_default);
     NurbsCurve(vector<Vector3d>, int k=k_default);
     NurbsCurve(string command, int n=n_default, int k=k_default);
-    NurbsCurve(double (*)(double), bool polar_coordinates=false, int n=n_default, int k=k_default);
-    NurbsCurve friend operator*(double, NurbsCurve);
-    NurbsCurve operator*(double);
-    Matrix getPoint(double t);
-    double get_min();
-    double get_max();
-    double get_t_from_height(double);
+    NurbsCurve(float (*)(float), bool polar_coordinates=false, int n=n_default, int k=k_default);
+    NurbsCurve friend operator*(float, NurbsCurve);
+    NurbsCurve operator*(float);
+    Matrix getPoint(float t);
+    float get_min();
+    float get_max();
+    float get_t_from_height(float);
     int get_n();
 };
 
@@ -181,12 +181,12 @@ class NurbsSurface{
     vector<Patch> patchList;
     NurbsSurface(int =m_default, int =n_default, string="", int =k_default);
     void cylinder_surface(NurbsCurve, NurbsCurve, NurbsCurve, Vector3d guide_direction=Vector3d(0, 0, 1));
-    void rocket_body(NurbsCurve, NurbsCurve, double (*)(double, double, double));
+    void rocket_body(NurbsCurve, NurbsCurve, float (*)(float, float, float));
     void binary_expand(NurbsCurve, NurbsCurve);
-    void water_expand(NurbsCurve, NurbsCurve, double, double, double);
+    void water_expand(NurbsCurve, NurbsCurve, float, float, float);
     void build_patches();
-    Matrix getPoint(double u, double v);
-    Matrix intersect(Ray view, Vector u_range=Vector(vector<double>({0, 1})), Vector v_range=Vector(vector<double>({0, 1})));
+    Matrix getPoint(float u, float v);
+    Matrix intersect(Ray view, Vector u_range=Vector(vector<float>({0, 1})), Vector v_range=Vector(vector<float>({0, 1})));
 };
 
 class Patch{
@@ -195,7 +195,7 @@ class Patch{
         Vector u_range, v_range;
         Box bBox;
         Vector3d centroid, norm;
-        Patch(NurbsSurface*, double, double, double, double);
+        Patch(NurbsSurface*, float, float, float, float);
         Box& getBox();
         Vector3d& getCentroid();
         bool intersect(Ray view, IntersectionInfo<Patch>*);
