+++ matrix.cpp
@@ -14,25 +14,25 @@
 
 using namespace std;
 
-static double Max=1e10, zero_threshold=1e-5;
+static float Max=1e10, zero_threshold=1e-5;
 
-bool negligible(double d, double th){
+bool negligible(float d, float th){
     if (th<0) th=zero_threshold;
     return abs(d)<=th;
 }
-bool approx(double a, double b, double th){
+bool approx(float a, float b, float th){
     if (th<0) th=zero_threshold;
     return abs(a-b)<=th;
 }
-double frand(){
-    return rand() % 10000 / (double)9999;
+float frand(){
+    return rand() % 10000 / (float)9999;
 }
 void strip(string& s, char c){
     s = s.substr(s.find(c) + 1);
 }
 
-Matrix::Matrix(int m, int n, double d):m(m), n(n), M(m, vector<double>(n, d)){ }
-Matrix::Matrix(double* input, int m, int n): Matrix(m, n, 0.0){
+Matrix::Matrix(int m, int n, float d):m(m), n(n), M(m, vector<float>(n, d)){ }
+Matrix::Matrix(float* input, int m, int n): Matrix(m, n, 0.0){
     for (int i=0; i!=m; i++)
         for (int j=0; j!=n; j++)
             M[i][j]=input[i*n+j];
@@ -47,10 +47,10 @@ Matrix::Matrix(vector<Vector3d> input): Matrix(input.size(), 3, 0.0){
 void Matrix::switch_row(int x, int y){
     swap(M[x], M[y]);
 }
-void Matrix::reset(double value){
+void Matrix::reset(float value){
     (*this) = Matrix(m, n, value);
 }
-void Matrix::random(double value){
+void Matrix::random(float value){
     srand(time(NULL));
     for (int i=0; i!=m; i++)
         for (int j=0; j!=n; j++)
@@ -86,7 +86,7 @@ bool Matrix::is_singular() const{
 bool Matrix::is_vector() const{
     return m==1;
 }
-double Matrix::mode(string type) const{
+float Matrix::mode(string type) const{
     if (type=="2"){
         Matrix cp = *this;
         return sqrt(dot(cp));
@@ -96,24 +96,24 @@ double Matrix::mode(string type) const{
     }
     else return 1;
 }
-double Matrix::norm(string type) const{
+float Matrix::norm(string type) const{
     return mode(type);
 }
 Matrix Matrix::normalized(string type) const{
     return (*this) / mode(type);
 }
-double Matrix::cond(string type) const{
+float Matrix::cond(string type) const{
     return (*this).inverse().mode(type)*(*this).mode(type);
 }
-double Matrix::get_min() const{
-    double output = 1e7;
+float Matrix::get_min() const{
+    float output = 1e7;
     for (int i=0; i!=m; i++)
         for (int j=0; j!=n; j++)
             output = std::min(output, M[i][j]);
     return output;
 }
-double Matrix::get_max() const{
-    double output = -1e7;
+float Matrix::get_max() const{
+    float output = -1e7;
     for (int i=0; i!=m; i++)
         for (int j=0; j!=n; j++)
             output = std::max(output, M[i][j]);
@@ -126,9 +126,9 @@ bool exist(int* choice, int i, int n){
         if (choice[j]==choice[i]) return true;
     return false;
 }
-double Matrix::determinant() const{
+float Matrix::determinant() const{
     int counter=0, i=0;
-    double sum=0., element;
+    float sum=0., element;
     int choice[m];
     for (int j=0; j!=m; j++)
         choice[j]=-1;
@@ -154,13 +154,13 @@ double Matrix::determinant() const{
     }
     return sum;
 }
-double Matrix::main_eigenvalue(double threshold, int* counter){
-    double last=-1.;
+float Matrix::main_eigenvalue(float threshold, int* counter){
+    float last=-1.;
     bool finished=false;
     Matrix x(m,1,1.);
     while (!finished){
         x = x * (*this);
-        double now=x.mode("infty");
+        float now=x.mode("infty");
         if (abs(now-last)<threshold)
             finished=true;
         x=x/now;
@@ -173,7 +173,7 @@ double Matrix::main_eigenvalue(double threshold, int* counter){
 
 Matrix Matrix::inverse2() const{
     Matrix ret = Matrix(2, 2, 0.0);
-    double denom = M[0][0]*M[1][1]-M[0][1]*M[1][0];
+    float denom = M[0][0]*M[1][1]-M[0][1]*M[1][0];
     ret[0][0] = M[1][1] / denom;
     ret[0][1] = -M[0][1] / denom;
     ret[1][0] = -M[1][0] / denom;
@@ -182,8 +182,8 @@ Matrix Matrix::inverse2() const{
 }
 Matrix Matrix::inverse() const{
     if (m==2 && n==2){
-        double denom = 1/(M[0][0]*M[1][1]-M[0][1]*M[1][0]);
-        double value[4] = { M[1][1] * denom, -M[0][1] * denom,
+        float denom = 1/(M[0][0]*M[1][1]-M[0][1]*M[1][0]);
+        float value[4] = { M[1][1] * denom, -M[0][1] * denom,
                 -M[1][0] * denom, M[0][0] * denom };
         return Matrix(value, 2, 2);
     }
@@ -212,8 +212,8 @@ Matrix Matrix::inverse() const{
     }
     return result;
 }
-double Matrix::dot(Matrix a) const{
-    double sum=0.;
+float Matrix::dot(Matrix a) const{
+    float sum=0.;
     for (int i=0; i!=m; i++)
         for (int j=0; j!=n; j++)
             sum+=a[i][j]*M[i][j];
@@ -243,14 +243,14 @@ Matrix Matrix::operator*(Matrix A) const{
                 result[i][j]+=M[i][k]*A[k][j];
     return result;
 }
-Matrix Matrix::operator*(double k) const{
+Matrix Matrix::operator*(float k) const{
     Matrix result = *this;
     for (int i=0; i!=m; i++)
         for (int j=0; j!=n; j++)
             result.M[i][j] *= k;
     return result;
 }
-Matrix Matrix::operator/(double k) const{
+Matrix Matrix::operator/(float k) const{
     Matrix result = *this;
     for (int i=0; i!=m; i++)
         for (int j=0; j!=n; j++)
@@ -276,7 +276,7 @@ void Matrix::operator+=(Matrix A){
         for (int j=0; j!=n; j++)
             M[i][j] += A.M[i][j];
 }
-double Matrix::operator^(Matrix& A) const{
+float Matrix::operator^(Matrix& A) const{
     return this->dot(A);
 }
 bool Matrix::operator==(Matrix& A) const{
@@ -288,7 +288,7 @@ bool Matrix::operator==(Matrix& A) const{
                 return false;
     return true;
 }
-Matrix operator*(double k, Matrix from){
+Matrix operator*(float k, Matrix from){
     Matrix result(from.m,from.n,0.0);
     int x=from.m, y=from.n;
     for (int i=0; i!=x; i++)
@@ -296,7 +296,7 @@ Matrix operator*(double k, Matrix from){
             result[i][j]=from[i][j]*k;
     return result;
 }
-vector<double>& Matrix::operator[](int i){
+vector<float>& Matrix::operator[](int i){
     return M[i];
 }
 ostream &operator<<(ostream& stream, const Matrix& mat){
@@ -359,35 +359,35 @@ Matrix Matrix::Identity(int n){
  *                  Vector Zone
  *-------------------------------------------------------*/
 
-Vector::Vector(int m, double init): Matrix(1, m, init){
+Vector::Vector(int m, float init): Matrix(1, m, init){
 }
 Vector::Vector(const Matrix A): Matrix(A){
     if (A.n == 1)
         (*this) = A.transpose();
 }
-Vector::Vector(const vector<double>& array): Matrix(1, array.size(), 0){
+Vector::Vector(const vector<float>& array): Matrix(1, array.size(), 0){
     for (int i=0; i!=n; i++)
         M[0][i] = array[i];
 }
 Vector3d Vector::to_3d(){
     return (Vector3d)(*this);
 }
-double& Vector::operator[](int i){
+float& Vector::operator[](int i){
     return M[0][i];
 }
-Vector operator*(double c, Vector v){
+Vector operator*(float c, Vector v){
     return c * v;
 }
 Matrix Vector::transpose(){
     return ((Matrix)(*this));
 }
-Vector Vector::min(Vector& v, double c){
+Vector Vector::min(Vector& v, float c){
     Vector output(v.n);
     for (int i=0; i!=v.n; i++)
         output[i] = std::min(v[i], c);
     return output;
 }
-Vector Vector::max(Vector& v, double c){
+Vector Vector::max(Vector& v, float c){
     Vector output(v.n);
     for (int i=0; i!=v.n; i++)
         output[i] = std::max(v[i], c);
@@ -415,9 +415,9 @@ Matrix operator*(Vector V, Matrix M){
 /* -------------------- Vector3d Zone ------------------------*/
 
 /*
-#define loadps(mem)		_mm_load_ps((const double * const)(mem))
-#define storess(ss,mem)	_mm_store_ss((double * const)(mem),(ss))
-#define storeps(ss,mem) _mm_store_ps((double * const)(mem), (ss))
+#define loadps(mem)		_mm_load_ps((const float * const)(mem))
+#define storess(ss,mem)	_mm_store_ss((float * const)(mem),(ss))
+#define storeps(ss,mem) _mm_store_ps((float * const)(mem), (ss))
 #define minss			_mm_min_ss
 #define maxss			_mm_max_ss
 #define minps			_mm_min_ps
@@ -429,11 +429,11 @@ Matrix operator*(Vector V, Matrix M){
 #define rotatelps(ps)		_mm_shuffle_ps((ps),(ps), 0x39)
 #define muxhps(low,high)	_mm_movehl_ps((low),(high))
 
-static const double flt_plus_inf = -logf(0);
-static const double __attribute__((aligned(16)))
+static const float flt_plus_inf = -logf(0);
+static const float __attribute__((aligned(16)))
   ps_cst_plus_inf[4] = {  flt_plus_inf,  flt_plus_inf,  flt_plus_inf,  flt_plus_inf },
   ps_cst_minus_inf[4] = { -flt_plus_inf, -flt_plus_inf, -flt_plus_inf, -flt_plus_inf };
-Vector3d::Vector3d(double x, double y, double z): Vector(3, 0.0){
+Vector3d::Vector3d(float x, float y, float z): Vector(3, 0.0){
     M[0].reserve(4);
     const __m128
         temp = _mm_setr_ps(x, y, z, 0);
@@ -442,7 +442,7 @@ Vector3d::Vector3d(double x, double y, double z): Vector(3, 0.0){
 Vector3d::Vector3d(const Matrix A): Vector(A){
     M[0].reserve(4);
 }
-Vector3d::Vector3d(const vector<double> array): Vector(3, 0.0){
+Vector3d::Vector3d(const vector<float> array): Vector(3, 0.0){
     M[0].reserve(4);
     storeps(loadps(&array[0]), &M[0][0]);
 }
@@ -471,7 +471,7 @@ Vector3d Vector3d::project(Vector3d& x, Vector3d& y, Vector3d& z){
     storeps(addps(ps, pz), &ret[0]);
     return ret;
 }
-double Vector3d::norm(){
+float Vector3d::norm(){
     const __m128 v = loadps(&M[0][0]);
     return _mm_cvtss_f32(_mm_sqrt_ss(_mm_dp_ps(v, v, 0x71))); 
 }
@@ -505,11 +505,11 @@ Vector3d Vector3d::operator+(Vector3d v) const{
     storeps(s, &ret[0]);
     return ret;
 }
-double Vector3d::operator^(Vector3d v) const{
+float Vector3d::operator^(Vector3d v) const{
     return _mm_cvtss_f32(_mm_dp_ps(loadps(&v[0]),
                 loadps(&M[0][0]), 0x71)); 
 }
-Vector3d Vector3d::operator/(double c) const{
+Vector3d Vector3d::operator/(float c) const{
     Vector3d ret;
     const __m128
         v1 = loadps(&M[0][0]),
@@ -518,7 +518,7 @@ Vector3d Vector3d::operator/(double c) const{
     storeps(s, &ret[0]);
     return ret;
 }
-Vector3d Vector3d::operator*(double c) const{
+Vector3d Vector3d::operator*(float c) const{
     Vector3d ret;
     const __m128
         v1 = loadps(&M[0][0]),
@@ -534,12 +534,12 @@ void Vector3d::operator+=(Vector3d v){
         s = addps(v1, v2);
     storeps(s, &M[0][0]);
 };
-Vector3d operator*(double c, Vector3d v){
+Vector3d operator*(float c, Vector3d v){
     return Vector3d(c*v[0], c*v[1], c*v[2]);
 };
 */
 
-Vector3d::Vector3d(double x, double y, double z):x(x), y(y), z(z){
+Vector3d::Vector3d(float x, float y, float z):x(x), y(y), z(z){
 }
 Vector3d::Vector3d(Matrix A){
     if (A.n == 1)
@@ -547,7 +547,7 @@ Vector3d::Vector3d(Matrix A){
     else
         *this = Vector3d(A[0][0], A[0][1], A[0][2]);
 }
-Vector3d::Vector3d(const vector<double> array): Vector3d(array[0], array[1], array[2]){
+Vector3d::Vector3d(const vector<float> array): Vector3d(array[0], array[1], array[2]){
 }
 void Vector3d::print(){
     cout<<"Vector 3d: "<<x<<", "<<y<<", "<<z<<endl;
@@ -574,16 +574,16 @@ Vector3d Vector3d::get_FinVector(Vector3d guide){
         return (z - z.dot(norm) * norm).normalized();
     }
 }
-double Vector3d::norm(){
+float Vector3d::norm(){
     return sqrt(x*x + y*y + z*z);
 }
 Vector3d Vector3d::normalized(){
     return (*this)/sqrt(x*x+y*y+z*z);
 }
-double Vector3d::get_min(){
+float Vector3d::get_min(){
     return std::min(std::min(x, y), z);
 }
-double Vector3d::get_max(){
+float Vector3d::get_max(){
     return std::max(std::max(x, y), z);
 }
 Vector3d Vector3d::operator-(Vector3d v) const{
@@ -598,22 +598,22 @@ Vector3d Vector3d::operator*(Vector3d v) const{
 Vector3d Vector3d::operator/(Vector3d v) const{
     return Vector3d(x / v.x, y / v.y, z / v.z);
 }
-double Vector3d::operator^(Vector3d v) const{
+float Vector3d::operator^(Vector3d v) const{
     return x * v.x + y * v.y + z * v.z;
 }
 Vector3d Vector3d::operator%(Vector3d v) const{
     return cross(v);
 }
-Vector3d Vector3d::operator/(double c) const{
+Vector3d Vector3d::operator/(float c) const{
     return Vector3d(x / c, y / c, z / c);
 }
-Vector3d Vector3d::operator*(double c) const{
+Vector3d Vector3d::operator*(float c) const{
     return Vector3d(x * c, y * c, z * c);
 }
-Vector3d Vector3d::operator+(double c) const{
+Vector3d Vector3d::operator+(float c) const{
     return Vector3d(x + c, y + c, z + c);
 }
-Vector3d Vector3d::operator-(double c) const{
+Vector3d Vector3d::operator-(float c) const{
     return Vector3d(x - c, y - c, z - c);
 }
 void Vector3d::operator+=(Vector3d v){
@@ -621,17 +621,17 @@ void Vector3d::operator+=(Vector3d v){
     y += v.y;
     z += v.z;
 }
-Vector3d operator*(double c, Vector3d v){
+Vector3d operator*(float c, Vector3d v){
     return Vector3d(c*v.x, c*v.y, c*v.z);
 }
-double& Vector3d::operator[](int i){
+float& Vector3d::operator[](int i){
     switch (i){
         case 0: return x;
         case 1: return y;
         case 2: return z;
     }
 }
-double Vector3d::dot(Vector3d v) const{
+float Vector3d::dot(Vector3d v) const{
     return x*v.x + y*v.y + z*v.z;
 }
 ostream& operator<<(ostream& stream, const Vector3d& v){
@@ -643,28 +643,28 @@ Vector3d Vector3d::min(Vector3d u, Vector3d v){
 Vector3d Vector3d::max(Vector3d u, Vector3d v){
     return Vector3d(std::max(u.x, v.x), std::max(u.y, v.y), std::max(u.z, v.z));
 }
-Vector3d Vector3d::min(Vector3d u, double c){
+Vector3d Vector3d::min(Vector3d u, float c){
     return Vector3d(std::min(u.x, c), std::min(u.y, c), std::min(u.z, c));
 }
-Vector3d Vector3d::max(Vector3d u, double c){
+Vector3d Vector3d::max(Vector3d u, float c){
     return Vector3d(std::max(u.x, c), std::max(u.y, c), std::max(u.z, c));
 }
-double Vector3d::get_x() const{
+float Vector3d::get_x() const{
     return x;
 }
-double Vector3d::get_y() const{
+float Vector3d::get_y() const{
     return y;
 }
-double Vector3d::get_z() const{
+float Vector3d::get_z() const{
     return z;
 }
-double dec(double r){
-    double pi = 3.14159265358979;
+float dec(float r){
+    float pi = 3.14159265358979;
     return 0.5*(cos(r*pi)+1);
 }
-Vector3d Vector3d::proceed(double r){
-    double xp, yp, zp;
-    double f = 1.0/3;
+Vector3d Vector3d::proceed(float r){
+    float xp, yp, zp;
+    float f = 1.0/3;
     if (r<=f){
         xp = dec(3*r)*x + (1-dec(3*r))*z;
         yp = dec(3*r)*y + (1-dec(3*r))*x;
